최소한의 샘플 코드로 Compose View에서 Swift 언어 표현하기
