+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.J2KB7thUniconKEEZAndroid"
         tools:targetApi="31">
         <activity android:name="com.j2kb.keez.view.home.MainActivity"/>
-        <activity android:name="com.j2kb.keez.view.markdown_test.MarkdownTestActivity"
+        <activity android:name="com.j2kb.keez.view.codeblock.CodeBlockActivity"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"/>
         <activity android:name=".view.login.LoginActivity"
             android:exported="true">
