+++ app/src/main/java/com/j2kb/keez/view/codeblock/CodeBlock.kt
@@ -1,0 +1,9 @@
+package com.j2kb.keez.view.codeblock
+
+object CodeBlock {
+    enum class Language {
+        JAVA, KOTLIN, SWIFT
+    }
+
+    data class Code(val language: Language, val sourceCode: String)
+}
