--- app/src/main/java/com/j2kb/keez/view/markdown_test/MarkwonSample.kt
@@ -1,16 +1,0 @@
-package com.j2kb.keez.view.markdown_test
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-abstract class MarkwonSample {
-
-    fun createView(inflater: LayoutInflater, container: ViewGroup?): View {
-        return inflater.inflate(layoutResId, container, false)
-    }
-
-    abstract fun onViewCreated(view: View)
-
-    protected abstract val layoutResId: Int
-}
