--- app/src/main/java/com/j2kb/keez/view/markdown_test/Tag.java
@@ -1,39 +1,0 @@
-package com.j2kb.keez.view.markdown_test;
-
-public enum Tag {
-    basics,
-    toast,
-    hack,
-    parsing,
-    block,
-    movementMethod,
-    links,
-    plugin,
-    recyclerView,
-    paragraph,
-    rendering,
-    style,
-    theme,
-    image,
-    newLine,
-    softBreak,
-    defaults,
-    spacing,
-    padding,
-    heading,
-    anchor,
-    lists,
-    extension,
-    textAddedListener,
-    editor,
-    span,
-    svg,
-    gif,
-    inline,
-    html,
-    knownBug,
-    precomputedText,
-    cache,
-    spanFactory,
-    reddit
-}
